+++ pygal/graph/base.py
@@ -22,7 +22,7 @@ Base for pygal charts
 """
 
 from __future__ import division
-from pygal._compat import u, is_list_like, to_unicode
+from pygal._compat import is_list_like
 from pygal.view import Margin, Box
 from pygal.config import Config
 from pygal.state import State
@@ -34,12 +34,11 @@ from pygal.adapters import (
     not_zero, positive, decimal_to_float)
 from functools import reduce
 from uuid import uuid4
-import io
 import os
 
 
 class BaseGraph(object):
-    """Graphs commons"""
+    """Chart internal behaviour related functions"""
 
     _adapters = []
 
@@ -72,18 +71,6 @@ class BaseGraph(object):
         ) is None or name not in self.state.__dict__:
             return super(BaseGraph, self).__getattribute__(name)
         return getattr(self.state, name)
-
-    def add(self, title, values, **kwargs):
-        """Add a serie to this graph"""
-        if not is_list_like(values) and not isinstance(values, dict):
-            values = [values]
-        if kwargs.get('secondary', False):
-            self.raw_series2.append((title, values, kwargs))
-        else:
-            self.raw_series.append((title, values, kwargs))
-
-    def add_xml_filter(self, callback):
-        self.xml_filters.append(callback)
 
     def prepare_values(self, raw, offset=0):
         """Prepare the values to start with sane values"""
@@ -230,107 +217,6 @@ class BaseGraph(object):
 
         del self.state
         self.state = None
-
-    def render(self, is_unicode=False, **kwargs):
-        """Render the graph, and return the svg string"""
-        self.setup(**kwargs)
-        svg = self.svg.render(
-            is_unicode=is_unicode, pretty_print=self.pretty_print)
-        self.teardown()
-        return svg
-
-    def render_tree(self):
-        """Render the graph, and return (l)xml etree"""
-        self.setup()
-        svg = self.svg.root
-        for f in self.xml_filters:
-            svg = f(svg)
-        self.teardown()
-        return svg
-
-    def render_table(self, **kwargs):
-        # Import here to avoid lxml import
-        try:
-            from pygal.table import Table
-        except ImportError:
-            raise ImportError('You must install lxml to use render table')
-        return Table(self).render(**kwargs)
-
-    def render_pyquery(self):
-        """Render the graph, and return a pyquery wrapped tree"""
-        from pyquery import PyQuery as pq
-        return pq(self.render(), parser='html')
-
-    def render_in_browser(self, **kwargs):
-        """Render the graph, open it in your browser with black magic"""
-        try:
-            from lxml.html import open_in_browser
-        except ImportError:
-            raise ImportError('You must install lxml to use render in browser')
-        open_in_browser(self.render_tree(**kwargs), encoding='utf-8')
-
-    def render_response(self, **kwargs):
-        """Render the graph, and return a Flask response"""
-        from flask import Response
-        return Response(self.render(**kwargs), mimetype='image/svg+xml')
-
-    def render_django_response(self, **kwargs):
-        """Render the graph, and return a Django response"""
-        from django.http import HttpResponse
-        return HttpResponse(
-            self.render(**kwargs), content_type='image/svg+xml')
-
-    def render_to_file(self, filename, **kwargs):
-        """Render the graph, and write it to filename"""
-        with io.open(filename, 'w', encoding='utf-8') as f:
-            f.write(self.render(is_unicode=True, **kwargs))
-
-    def render_to_png(self, filename=None, dpi=72, **kwargs):
-        """Render the graph, convert it to png and write it to filename"""
-        import cairosvg
-        return cairosvg.svg2png(
-            bytestring=self.render(**kwargs), write_to=filename, dpi=dpi)
-
-    def render_sparktext(self, relative_to=None):
-        """Make a mini text sparkline from chart"""
-        bars = u('▁▂▃▄▅▆▇█')
-        if len(self.raw_series) == 0:
-            return u('')
-        values = list(self.raw_series[0][1])
-        if len(values) == 0:
-            return u('')
-
-        chart = u('')
-        values = list(map(lambda x: max(x, 0), values))
-
-        vmax = max(values)
-        if relative_to is None:
-            relative_to = min(values)
-
-        if (vmax - relative_to) == 0:
-            chart = bars[0] * len(values)
-            return chart
-
-        divisions = len(bars) - 1
-        for value in values:
-            chart += bars[int(divisions *
-                              (value - relative_to) / (vmax - relative_to))]
-        return chart
-
-    def render_sparkline(self, **kwargs):
-        spark_options = dict(
-            width=200,
-            height=50,
-            show_dots=False,
-            show_legend=False,
-            show_x_labels=False,
-            show_y_labels=False,
-            spacing=0,
-            margin=5,
-            explicit_size=True
-        )
-        spark_options.update(kwargs)
-        return self.render(**spark_options)
 
     def _repr_svg_(self):
         """Display svg in IPython notebook"""