+++ pygal/graph/histogram.py
@@ -33,7 +33,6 @@ class Histogram(Graph):
     _dual = True
     _series_margin = 0
 
-
     @cached_property
     def _secondary_values(self):
         """Getter for secondary series values (flattened)"""
@@ -128,9 +127,11 @@ class Histogram(Graph):
             self._box.ymin, self._box.ymax = ymin, ymax
 
         x_pos = compute_scale(
-            self._box.xmin, self._box.xmax, self.logarithmic, self.order_min)
+            self._box.xmin, self._box.xmax, self.logarithmic, self.order_min
+        ) if not self.x_labels else list(map(float, self.x_labels))
         y_pos = compute_scale(
-            self._box.ymin, self._box.ymax, self.logarithmic, self.order_min)
+            self._box.ymin, self._box.ymax, self.logarithmic, self.order_min
+        ) if not self.y_labels else list(map(float, self.y_labels))
 
         self._x_labels = list(zip(map(self._format, x_pos), x_pos))
         self._y_labels = list(zip(map(self._format, y_pos), y_pos))